pyproject.toml
[tool.poetry]
name = "pizza-bake"
version = "0.1.0"
description = ""
authors = ["Ananthu Krishnan <dev.ananthu.krishnan@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.2"
django-cors-headers = "^4.3.1"
django-rest-framework = "^0.1.0"
pillow = "^10.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
